+++ src/main/java/com/galaxyfreedom/introduction/config/PaginationConfig.java
@@ -1,0 +1,14 @@
+package com.galaxyfreedom.introduction.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.web.config.PageableHandlerMethodArgumentResolverCustomizer;
+
+@Configuration
+public class PaginationConfig {
+
+    @Bean
+    public PageableHandlerMethodArgumentResolverCustomizer paginationCustomizer() {
+        return resolver -> resolver.setOneIndexedParameters(true);
+    }
+}
