+++ src/main/java/com/galaxyfreedom/introduction/profile/repository/ProjectRepository.java
@@ -1,6 +1,8 @@
 package com.galaxyfreedom.introduction.profile.repository;
 
 import com.galaxyfreedom.introduction.profile.entity.Project;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,5 +11,5 @@ import java.util.UUID;
 
 @Repository
 public interface ProjectRepository extends JpaRepository<Project, UUID> {
-    Set<Project> findAllByProfile_Id(UUID profileId);
+    Page<Project> findAllByProfile_Id(UUID profileId, Pageable pageable);
 }
