+++ src/main/java/com/galaxyfreedom/introduction/profile/service/ProjectService.java
@@ -28,7 +28,7 @@ public class ProjectService {
                 .orElseThrow(() -> new EntityNotFoundException("Project with id " + id + " not found"));
     }
 
-    public Set<Project> findAllByProfile_Id(UUID profileId) {
-        return projectRepository.findAllByProfile_Id(profileId);
+    public Page<Project> findAllByProfile_Id(UUID profileId, Pageable pageable) {
+        return projectRepository.findAllByProfile_Id(profileId, pageable);
     }
 }
